+++ ChangeLog.md
@@ -13,6 +13,8 @@ and this project adheres to
 
 ### Added
 
+- support for _Gitpod VS Code_
+
 - color mode switching to the REPL. The REPL commands `!lightmode` and
   `!darkmode` switch Venice's color mode to adapt to the REPL's 
   terminal color mode.